+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Added support to ```Pagination``` library
 - Added support to ```Form Validation``` library
 - Added support to ```Input``` library
+- Added support to ```Image Manipulation``` library
 
 ## [0.0.7]
 ### Added